Docker compose separate folder for vendor per php version
