+++ docker-compose.yml
@@ -26,15 +26,28 @@ x-aliases:
         - DDAGENT_HOSTNAME=ddagent_integration
       cap_add:
         - SYS_PTRACE
+  - &base_54_service
+      <<: *base_php_service
+      volumes:
+        - .:/home/circleci/app
+        - ./tmp/docker/54/vendor:/home/circleci/app/vendor
+        - .composer:/home/circleci/.composer
+        - .scenarios.lock:/home/circleci/app/.scenarios.lock
+  - &base_56_service
+      <<: *base_php_service
+      volumes:
+        - .:/home/circleci/app
+        - ./tmp/docker/56/vendor:/home/circleci/app/vendor
+        - .composer:/home/circleci/.composer
+        - .scenarios.lock:/home/circleci/app/.scenarios.lock
 
 services:
   '5.4-debug': { <<: *base_php_service, build: 'dockerfiles/debug_php-5.4' }
   '5.4': { <<: *base_php_service, build: 'dockerfiles/php-5.4' }
 
-  '5.6': { <<: *base_php_service, image: 'datadog/docker-library:ddtrace_php_5_6' }
-  '5.6-zts': { <<: *base_php_service, image: 'circleci/php:5.6-zts' }
-  '5.6-debug': { <<: *base_php_service, build: 'dockerfiles/debug_php-5.6' }
-
+  '5.6': { <<: *base_56_service, image: 'datadog/docker-library:ddtrace_php_5_6' }
+  '5.6-zts': { <<: *base_56_service, image: 'circleci/php:5.6-zts' }
+  '5.6-debug': { <<: *base_56_service, build: 'dockerfiles/debug_php-5.6' }
   '7.0': { <<: *base_php_service, image: 'datadog/docker-library:ddtrace_php_7_0' }
   '7.1': { <<: *base_php_service, image: 'datadog/docker-library:ddtrace_php_7_1' }
   'ubuntu-16.04-7.1': { <<: *base_php_service, build: 'dockerfiles/ubuntu-16.04-php-7.1' }
