+++ tests/OpenTracerUnit/ScopeManagerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DDTrace\Tests\Unit\OpenTracer;
+namespace DDTrace\Tests\OpenTracerUnit;
 
 use DDTrace\OpenTracer\ScopeManager;
 use DDTrace\OpenTracer\Tracer;
