+++ tests/OpenTracerUnit/TracerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DDTrace\Tests\Unit\OpenTracer;
+namespace DDTrace\Tests\OpenTracerUnit;
 
 use DDTrace\Configuration;
 use DDTrace\OpenTracer\SpanContext;
